[Review] Review of #1633 (#1669)
Squash and merging into 1633.
