+++ localnet/pocketd/config/relayminer_config_full_example.yaml
@@ -87,10 +87,12 @@ suppliers:
   # Example with different RPC type configurations for XRPL EVM
   - service_id: xrplevm
     listen_url: http://0.0.0.0:8545
+
     # Default service configuration (used when no specific RPC type matches)
     service_config:
       backend_url: http://xrplevm-node:8545
       forward_pocket_headers: true
+
     # RPC type specific configurations for different XRPL EVM endpoints
     rpc_type_service_configs:
       # JSON-RPC endpoints (Ethereum-compatible API)
@@ -99,6 +101,7 @@ suppliers:
         headers:
           Authorization: "Bearer xrpl-api-key"
         forward_pocket_headers: true
+
       # REST API endpoints (Cosmos SDK REST for XRPL sidechain)
       rest:
         backend_url: http://xrplevm-node:1317
@@ -106,10 +109,12 @@ suppliers:
           username: xrpluser
           password: xrplpass
         forward_pocket_headers: true
+
       # CometBFT RPC endpoints (for consensus layer queries)
       comet_bft:
         backend_url: http://xrplevm-node:26657
         forward_pocket_headers: true
+
       # WebSocket endpoints (for real-time event streaming)
       websocket:
         backend_url: ws://xrplevm-node:8546
